✨ feat: implement NavBreadcrumbs component and refactor DashboardNav and AppSidebar to use HydrateClient
